requirements.txt
streamlit>=1.36.0
google-generativeai>=0.3.2
python-dotenv>=1.0.1
nbformat>=5.7.0
PyPDF2>=3.0.0
pandas>=2.0.0
pydot>=1.4.2
graphviz>=0.20.3
